Refactor system API by consolidating endpoints and models
- Moved system API endpoints and data models from `system.py` to `__init__.py` for better organization.
- Retained functionality for retrieving system information, health checks, and statistics.
- Enhanced code structure by centralizing related components within the same module.
